MB-58119 Do not decrement NumReplica2 for drop_replica action
For alter index with "drop_replica" action, the expectation is to
repair the dropped replica during next rebalance. In the current
implementation, this does not happen as the NumReplica2 counter
is decremented when a specific replica is being dropped.

When repair has to be performed, metadata_provider validates the
replica count from NumReplica2 value. If the current replica in
cluster are same is the value in NumReplica2, the repair is skipped.

As the expectation is to repair the dropped replica, metadata_provider
should not update the NumReplica2 counter when a specific replica is
being dropped from the cluster.

Change-Id: I0efa75d2508650e41ca0a0743fff6d85264c181c