MB-57378 Populate siblingIndex for shard replica repair
Sibling index will only be populated when an entire shard is
undergoing repair. Rebalancer is expected to copy the shard
from the node on which replica shard exists and trigger repair.

Due to variation in the number of replicas created, it is possible
that the replica shard may not contain all the indexes required
for repair. In such a case, the remaining set of index should be
repaired via DCP

Change-Id: Ia337a7d22d7dc30e605392aada83d1fce361316b