MB-57802 Group indexes based on alternate shardId
Indexer will group all index usages based on alternate shardId.
A proxy index usage will be generated with the load stats
equivalent to union of load stats of individual indexes.

The proxy index will be present as single large index to planner
for further planning purposes. The defnId of the proxy instance
will be based on slotId (derived from alternateID) so that planner
will not move the proxy IndexUsage to a node that already has same
slotId and different replica number. The partnId of a proxy is always
"0" - A proxy index will be treated as a non-partitioned index with
multiple replicas.

As instances get shared between shards with same slotID, during
rebalance, if planner decides to move the proxy to a node that
contains same slotID, a replica violation will be generated and
planner can choose to skip the movement.

The initialNode of proxy usage is determined based on the initialNode
values of individual indexes. If atleast one node has non-nil initialNode
value, then it means that a node exists with alternateShardId relevant
to the proxy. In such a case, populate the initialNode of the proxy with
the initialNode of the index.

If all indexes have nil initialNode, it means that the entire shard has
to be replica repaired.

Change-Id: I87257706d6e54eb72be363f1220dc19f6946f17e