MB-57378 Add replica repair support for file transfer rebalance
Replica repair for file transfer rebalance happens on following rules:

a. Each partition of an index can go to different slot

b. Instances with same partition Id of replica instances share the same
   slotId.
   E.g, partnId: 0 for replica: 0 and replica:1 share same slotId
	partnId: 1 for replica: 0 and replica:1 share same slotId

	partnId: 0 and partnId: 1 can have different slotIds i.e. each
	partition of an index can go to different slot

c. The replicaId of index instance and replica number of alterante
   shardId match i.e. partn-0 of replica: 1 will always go to a shard
   with replica number "1". Similarly, partn-0 of replica:2 will always
   go to a shard with replica number "2"

Once planner decides on the slot on which the index has to be repaired
goes to, it will place the index on the node in which the slot (with
corresponding replica number exists). Planner will override the
ExcludeIn flag behaviour in this case. If the index to be repaired
has a valid slot, then the index will not be added to the list of
eligibles so that the index will not be moved in isolation.

Planner can also place the index on a deleted node as the entire shard
can be moved along with the instance being repaired from the deleted
node during rebalance.

Change-Id: Id611e35e859429f4ea489d62997e154bf476ab19