MB-58279 Populate alternateShardIds for indexes getting upgraded (And re-plan placement if necessary)
Planner follows the rule that the replicas of a partition share
same slotId - The alternate shardId differs only by replica ID.

During rebalance, two cases can exist:

a. The index being moved to a 7.6+ node does not have alternate
shardIds

b. The index being moved to 7.6+ node has alternate shardIds

(a) can happen when the index is created on a <7.6 version of index
node (or) created via <7.6 version of query node. In either cases,
planner will check if there is any replica available for this
partition in the cluster. If such replica exists, then planner will
use the slot from the replica instance and build a new alternate
shardId.

E.g., if replica 1 of an index is being moved to a 7.6+ node, then
the slotId of replica 0 (say s0) will be taken and a alternate shardId
with the combination of s0 and replicaId: 1 (i.e. s0-1) will be
generated for index movement.

As a second step, planner will also check if such shard (s0-1) exists
on a different node (than the one the index is moving to). In such
a case, the index is forced to move to the node on which the shard
exists and then a replanning is performed. During re-plan, all
the indexes of the shard move as a single unit (i.e. as a proxy).

Case (b) can happen when rebalance runs with enableShardAffinity set
to false. This is disturb the shard affinity of the indexes. This is
currently not handled in the code (The support for the same will be
added in future patches)

Change-Id: I245c420580b34b134e98aaaac9f13393f6ab4fd0