MB-58279 Add per node settings endpoint to test upgrade cases
The per node settings are applicable only for the current node
and accept only one setting: "indexer.thisNodeOnly.ignoreAlternateShardIds"

This setting will reset the alternate shardIds in the index definition
at lifecycle manager. This will help to test upgrade cases where
indexes are created without alternate shardIds. The testcode required
to reset the alternateShardIds in index definition is compiled only
for CI machines and it is a no-op for production deployments.

A call to this setting endpoint is applicable only for this node
and all other index service nodes in the cluster are immune to the
changes made. The value is cached in memory and gets reset on indexer
restart.

Change-Id: Id9b98eb9ea068023f4f5717368e11c6985f3d7dd