MB-58399 Make binSize used for sorting shards configurable
Currently, planner uses bin based algorithm to sort shards
based on disk usage. This means that if 2 shards have 20G
and 21G of disk usage but the shard with 20G has more instances
than the shard with 21G, then the shard with 21G disk usage will
be picked up for placement.

This decision is taken by dividing the shards into bins -
The span of each bin is 2.5G by default (1% of maxDiskUsagePerShard).
Two shards in same bin have equal probability and the shard with less
number of instances will be picked up for placement.

This patch makes the bin size configurable so that placement
of an index on a shard can be controlled at runtime.

Change-Id: I695b4ebc2703ce3d52e0b040ae5196c20aeb724e