MB-58297 Retrieve client stats asynchronously and process client stats in bootstrap
This patch introduces two changes:

a. Process OPCODE_CLIENT_STATS in bootstrap - This will make sure that
client does not wait indexer to move to ready state to respond to the
call. During bootstrap, indexer will broadcast the full list of stats
to client as and when the list of indexes changes - Therefore, it is
fine for indexer to process OPCODE_CLIENT_STATS in bootstrap as the
next change in indexer list will be broadcasted to client and client
can update its list

b. Retrieve the client stats asynchronously. By retrieving client stats
asynchronously, watcher does not have to wait for stats from indexer.
This means that the indexes belonging to this node may not be picked
up for scans if replicas are available as client does not have the load
statistics of indexes on the node until getClientStats succeeds. All
indexes without replicas will be picked up for scans irrespective of
the stats values

In case, indexer broadcasts the full list of stats to watcher before
getClientStats succeed, then the compare-and-swap (CAS) in getClientStats
would fail as CAS would succeed only if clientStats are nil. This will
make sure that the full index stats broadcast always takes precedence
over updation via getClientStats.

Change-Id: I2cb9bb8b6c68626aef9feb427febd860be7d4435
(cherry picked from commit 051bacc60341419f3c13e2feca94a6a5af9b1078)