MB-58275 Add wait after rebalance cleanup for drop to finish
If a rebalance fails, then as part of cleanup, rebalancer will
drop indexes with in-progress builds on destination nodes and for ready
indexes on the source nodes.

Indexer returns success once the bookkeeping is updated with the deleted
index information. However, the actual index data cleanup is async.

If the next rebalance starts soon after, it is possible for the next
create index to get processed before the index data cleanup. If this
happens for the same index instance, there can be an index instance
without any storage file.

This patch adds a wait after a failed rebalance if a drop index was
issued to the indexer as part of cleanup.

With emptyNodeBatching, rebalancer uses large batch sizes for transfer.
In case of failure, this means more time may be required for async
cleanup of index data files. This patch adds wait only if
emptyNodeBatching has been enabled.

A better solution would be to add bookkeeping within indexer to keep a
track when async cleanup of indexes finish and this should be check by
the rebalancer before marking cleanup as finished. This will be taken up
in a separate patch.

Change-Id: I6213b9af719b89c61601bd2e1d9f42e9f54ef91a