MB-57932 close active connections to failing index node during scan
We track the active connections used for a scan at the connection pool
level and close all active connections when closing the GSIScanClient.
This will retry the scan on replicas if we have not exhausted the retry
limit during the current scan.

Changes: 1. Bookkeep and update active connections in the connection
pool
         2. After the pool is closed, new scans are terminated

Change-Id: I2fa4ad802b53533c405f7e2698f6d6d31e40035b