MB-57803 Fix the computation of AvgResidentPercent at planner
This patch introduces two changes for the computation of
AvgResidentPercent for proxy indexes:

a. Expose recs_in_mem and recs_on_disk stats for both
mainstore and backstore to planner.
These stats will be used to compute the records in memory
vs total records available for all the index instances
belonging to a shard. Based on this, the resident ratio of
the shard is determined

The stats are only available from >=7.6 version of the indexer

b. For <7.6 version of indexer, recs_in_mem and recs_on_disk are
not available. In such cases, planner estimates the records using
items_count.

At indexer, the combined_resident_percent stat is calculated based
on the formula: a * 100 / b where

a is (mainstore_recs_in_mem + backstore_recs_in_mem)
b is (mainstore_total_recs + backstore_total_recs)

To simplify the calculation, we can approximate mainstore_total_recs
and backstore_total_recs to items_count. Therefore, the records in
memory for each store of the index can be approximated as:
items_count * resident_percent. For secondary index, both mainstore
and backstore will be considered separately. So, the records in memory
will be approximated to 2 * items_count * resident_percent. For
primary index, only one store will be considered. So, it will be
approximated to just items_count * resident_percent

Change-Id: I463519c72a2d24e997b0b01be50b428d8daa4a1d