MB-58373 Limit the number of shards per node
Currently, the number of shards per node are decided based on
4% of plasma memory quota. If the indexer node is configured
with 100G quota, then the number of shards that can be created
will be equal to 3686. This will be in-efficient for plasma.

Instead of allowing to create large number of shards, cap the
number of shards per node to 200.

Change-Id: Ic7abfa39fce36979c1831ce028a6e3f174c45243