MB-57803 Update shard stats at indexer node
Planner will query the /stats/storage/shard endpoint
(available from 7.6 version) and updates the stats of
the proxy with the shard stats.

As some indexes can get added to the shard either due to
replica repair (or) due to upgrade cases, the stats of
those indexes will be added on top of existing shard stats.

Change-Id: Id87d22676c16399027f6b9da5fec131794122cde