MB-57379 Wait for real index to recover before processing proxy index
Consider a partitioned index placed on nodes n1 and n2. A new node
n3 is added to the cluster and both n1 and n2 are going out of the
cluster.

In such a case, two shard tokens get generated for the index movement
(One for each node). The first index to recover at lifecycle manager
will use the actual real instance Id of the index. All the other
partitions will be built using proxy and later merged.

If proxy is recovered before real index and merge is initiated, then
merge would fail as real index is still in async recovery. To avoid
such cases, proxy recovery will wait for real index recovery to
finish. This will ensure that proxy merge always succeeds.

Also, the recovered state of the deferred index is reset to CREATED
at the end of "handleInstRecoveryResponse" as READY state does not
matter for indexer. Till build statement is received, index should
stay in created state. However, index metadata is updated with READY
state

Change-Id: I7df6d0f449d6604a1b6852f1114368db7df84421