MB-58377 Handle EquivalentIndex check for proxy instances
For proxy instances, equivalent index check happens for
each instance in the proxy. Even if once index in the proxy
does not have an equivalent index in other proxy, then the
proxies are not considered equivalent.

It is possible that one proxy has more instances that the
other but still equivalent index check holds. E.g., if
idx_1, idx_2 belong to proxy_1 and idx_3 belong to proxy_2
and if idx_1, idx_2 and idx_3 are equivalent, then proxy_1
and proxy_2 are considered equivalent.

Change-Id: I44d6d866ffee257857ad3aa8e14cd74de1995a27