MB-58538 Handle failures during index recovery
If indexer receives any error while recovering from a shard,
currently indexer cleans up the instance and proceeds with
rebalance. This will lead to index getting dropped on source
while the index is not transferred to destination.

This patch addresses this issue by updating index metadata with
the error observed during recovery. Shard rebalance will observe
the error and then fail the rebalance. The index metadata is
cleaned up during the token cleanup phase of rebalance.

Change-Id: I53a8f40cb74af5bd7dc988d33d5f2561623175c7