MB-58509 Generate target nodes per partition
When populating alternate shardIds, targetNodes list i.e. the
list of nodes to which the new indexes are going to is generated
at index definition level.

For a partitioned index, it is possible that each partition can go
to a different slot. So, consider the following case:

a. Slot-1 exists on nodes n0, n1, n2, n3 with replicas 0,1,2,3 on
these nodes respectively

b. A new index is getting created with 2 replicas and 3 partitions
with partnId: 1 of the index going to nodes n2, n3 and partnId: 2
of the index going to n3, n4 and partnId: 3 going to n1, n4.
With the current logic, Slot-1 will be picked up for placing the new
index - but this is incorrect because replicaId: 0 of Slot-1 exists on
node n0 and n1 but there is no partition going to nodes n0 and n1.

The reason for this behaviour is because the target nodes are generated
at index definition level. So, planner considers all n0, n1, n2, n3 as
valid placement.

To avoid such incorrect groupings, the target nodes are now generated
per partition level. This will make sure that slot pruning will happen
correctly and will not lead to incorrect grouping.

Change-Id: Ie825a205c0b59e77ab0e0ef2623041ffcd1e675f