MB-58370 Use DCP to move indexes if destination is at a lower compat version
When planning to move indexes between two nodes, if the destination node
is at a lower shard compatibility version than source, then use DCP
to move tokens instead of shard movement as destination may not understand
the shard on source completely.

This scenario can occur when indexes are being moved from a higher version
node to a lower version node.

Change-Id: Ic72811a570aebcc4cc317bf5d3f524aad5cebaf8