MB-57807 Skip index drop on source for replica repair during cleanup
For replica repair case, skip dropping indexes on source node
when cleaning up transfer tokens.

Also, if a transfer token has moved to Ready state and there is no
siblingToken, then indexes can be dropped on source node even if
there is no DropOnSource token posted. This is because indexer
need not honour sub-cluster affinity when there is no siblingToken
and the transfer token in state ShardTokenReady means that all indexes
are caught up on destination node.

Therefore, it is safe to drop indexes on source node if sibling token
is empty.

Change-Id: I8a5361fd81ed1b44d01e34deba5e3835e83b6322