MB-58683 Re-name alternate shardId during shard repair
When a shard is being repaired, it will be copied from another
node in the cluster. The replica number of the alternate shardId
in shard metadata should match the replica ID of the instances
on the destination node.

Hence, rename the alternate shardID in shard metadata during
transfer so that on destination node, the shard will be recovered
will correct alternate shardId

Change-Id: Ic73ebb2843b76454bbdbd668576d970b123c80e0