MB-58378 Reset alternate shardIds during restore
When restoring indexes, the index grouping from the time of
backup may not be valid as there can be a change in cluster
topology or index topology. In such cases, performing a
regrouping would probably help to optimise on shard placement.

To handle such cases, the alternate shardIds in index metadata
will be reset at the time of restore based on the following rules:

a. Restore to empty cluster will not reset alternate shardIds

b. If the number of empty nodes in the cluster are less than the
nodes from which indexes are being restored, then alternate shardIds
would be rest

c. When restoring to an existing node in the cluster and the node already
has some indexes, then alternate shardIds will be reset.

Also, ungroup indexes after planning as buildIndexHostMapping expects
indexes to be ungrouped.

Change-Id: I01a1169c2d210e70e13e081da88ba6ae7359f395