MB-58382 Reset alternate shardIds in token defn during dcp rebalance
During DCP rebalance, if indexes are being moved out of a node, then
reset alternate shardIds in the index definition so that indexes with
same alternate shardIds do not exist on multiple nodes.

Reset of alternate shardIds happen only if planner is run as a part of
rebalance master during rebalance. When shard rebalancer invokes DCP
rebalancer, it does not run planner. So, the alternate shardIds in
index definition do not get reset.

Change-Id: I24bdd0236518a8a8d48d71989ab0305ded13f973