MB-58481 Enable empty node batching for shard transfer tokens
Empty node batching is enabled in the following cases:

a. Non-serverless deployment
b. Cluster is fully upgraded to post 7.6+
c. The config "indexer.rebalance.enableEmptyNodeBatching" is set to true
d. The destination node on which indexes are being built is empty
e. There is atleast one DCP token for the destination node

The main aim of empty node batching is to skip the destination
node from serving scans while DCP builds are happening as DCP
builds can saturate node resources and scan latencies can get
impacted. For shard rebalance alone, this is not a problem as
shard rebalance relies on shard copy which does not consume much
resources (2-4 cores based on the perf experiments).

However, if there are DCP tokens involved (either due to replica
repair (or) upgrade cases), then if a shard is moved to a destination
node, indexes on the shard start scans while there are DCP tokens yet
to be processed, then the DCP builds can saturate node resources and
impact scan latencies. Hence, shard rebalance will enable empty node
batching for the shard tokens only if empty node batching is enabled
and DCP transfer tokens are involved in the movement to destination node.

Change-Id: Ia897e2c876ffa8f7220f8324f27a5a544c0b02ed