MB-58481 Add support for empty node batching in shard rebalancer
Shard rebalancer will follow the below protocol for empty node
batching:

a. If IsEmptyNodeBatch is set to true in a transfer token, then
after recovering all indexes from the shard and catching up with
KV, IsPendingReady flag will be set to true

b. If IsEmptyNodeBatch is set to false, then the token is moved
to ShardTokenReady state

c. Once all tokens move to either ShardTokenDeleted state or
ShardTokenRecoverShard state with IsPendingReady set to true,
then master will initiate DCP rebalance

d. Once DCP rebalance is done, rebalance master will move the
transfer token state to ShardTokenMerge

e. Destination node would process this state and update the
RState of all the indexes on the node. Once this process is
complete, the transfer token is moved to ShardTokenReady state.
(Before moving the token state to ShardTokenReady, if there are
any DCP tokens that needs RState change, then shard rebalancer
would also move the RState of DCP tokens. This change is not
a part of this patch but it will be added eventually)

After a token is moved to ShardTokenReady state, normal rebalance
flow would continue.

Change-Id: I315d034e65449c6219eecd19f9891e9ae44c1632