MB-58481 Skip moving RState if shard rebalancer started DCP rebalancer
When shard rebalancer starts DCP rebalance, then the following cases
need to be handled:

a. An index is getting repaired via DCP and is going to dest. node: n1
with alternate shardIds: [x-0-0, x-0-1]

b. Shard rebalancer also decided to move the shards with alternate
shardIds [x-0-0, x-0-1] to node n1 as a part of shard rebalance

c. As per the protocol, first the shard is copied and then the index
will be repaired on the shard

d. Due to empty node batching, the RState of indexes on shard will not
be moved to RebalActive

e. If the index building via DCP finishes and changes the RState to
active, then the index will be serving scans

f. After DCP rebalance finishes, if changing RState for the indexes on
the shard fail, then the indexes will be cleaned up on destination.

This will lead to the case where the shard is cleaned up on destination
with a live index on it. To handle such cases, the DCP tokens should not
change the RState of the index until the RState of the indexes of the
shard are changed.

This patch adds the support in DCP rebalancer to skip changing RState
of the tokens and handling over the control back to shard rebalancer.

Change-Id: Ib46b2ab0f663dd72d445a35aa0c2e83d297856e4