MB-58920 Reset shardIdsForDest in index definition metadata
Shard rebalance may populate ShardIdsForDest but they should not
be populated in the meta under index definition as they are not a
part of index definition. E.g., a simple create index statement will
not have shardIdsForDest in index definition metadata. They can be
added under topology though (as tenant aware planner uses the information
from topology to group indexes for serverless deployments). Hence reset
shardIds when persisting definition in index metadata.

If the shardIds are not reset, then if the index is moving via DCP rebalance,
it will continue to carry the shardIds from old nodes and index initilisation
on new node can fail as shard Ids will not be available on new nodes.

Change-Id: I3d50ee4647d6070286f71e91d826a7a5b491ba8e