MB-59011 Broadcast full set of stats for 5 min after indexer restart
Two types of race conditions are causing CI to fail sporadically:

Case-1:

a. Indexer node failed over. The metadata information of the index is deleted
at the client

b. New indexer is added and replica repair is perfomed via shard repair. Before
indexer updates the stats of the index at client, indexer panics (as expected
by the test)

c. Due to indexer crash, the peerPipe connection between indexer and client is
disconnected. Watcher will take some finite time to connect back to indexer i.e.
after a back-off

d. While watcher is waiting in backoff, indexer broadcasts full list of indexes.
As watcher is not active, the notification will not be received by watcher

e. When watcher re-connects again, it syncs with peer and updates the load stats of
the index to defaults. As the notification from (d) is missed, the index stats will
not be updated until indexer does a full broadcast of stats again (which happens every
5 min). This is causing the replica to have stale items count and scans failing with
"zero requests" seen for replica.

Case-2:

Similar to Case-1 but happens with swap rebalance. Before stats for index on destination
are broadcasted, destination node encounteres a panic (as expected in the test). As a part
of cleanup, index on source node is dropped which will lead to cleanup of index in topology.
Now the load statistics of the index are stale in meta repo. Due to indexer panic, wathcer
is in backoff. If (d) and (e) steps from Case-1 happen again, then the indexer will live
with stale stats for 5 min.

The fix for this issue is to broadcast full list of stats for 5 min. after indexer restarts.
As maximum back-off time for watcher is 30 seconds and sync with peer usually does not take
more than a few seconds, broadcasting full set of stats for 5 min. ensures that client
is always upto date with indexer stats after indexer restarts.

With this fix, there should be no sporiadic failures in CI.

Change-Id: I741b63ae61044fc3f0cf301056f751c3e6e86983