MB-59052 Handle race between doFinish and Cancel()
The following sequence of events can lead to a race:

a. Rebalance cancel is called by ns_server

b. Rebalance follower invoked ShardRebalancer::Cancel

c. Rebalance master inititated cleanup and deleted the global token

d. Rebalance follower observed the token deletion before metaKV is
cancelled. This triggered finishRebalance()

e. Rebalance follower closed metaKV channel and waiting for all
go-routines to exit

f. finishRebalance will also invoke cleanupOnce.Do(doFinish) which
will also wait for all go-routines to exit

At this point, there are two go-routines doing wg.Wait() and golang will
panic. The fix for this issue is to exhaust the cleanupOnce.Do() in
Cancel() so that doFinish is not executed when Cancel is called.

Also, another race condition can happen with regard to DCP cancel

a. startDcpRebalance() is waiting for mu.Lock() to start DCP rebalance

b. Cancel() acquired the lock and saw that dcpRebr is nil

c. startDcpRebalance() will acquire the lock and start the rebalancer object
as dcpRebrOnce is still not used.

To avoid this race condition, dcpRebrOnce is exhausted if dcpRebr is found to be
nil when cancelling the DCP rebalance. The processRebalanceChange go-routine is
also added to wait group so that the go-routine exits gracefully when rebalance
is cancelled.

Also, dcpRebalance cancel is moved ahead of shard rebalance as shard rebalance
would wait for DCP rebalance to finish in processRebalanceChange()

Change-Id: I85d2b0173d7107261fa3385f5826eef117a204c3