MB-58480 fix race in settings update
bg: currently, on new indexer being added to the cluster, we run into a
race by the following events (leading to indexer restart):

1. a node with index service added to the cluster for the first time

2. indexer ready

3. in handleReady for settings manager, we see that we need to update
settings in metakv so we push new value to metakv for path
`IndexingSettingsMetaPath`;
this is due to change in memory_quota

4. again in handleReady, we now try to apply new settings by using
`IndexingSettingsShardAffinityMetaPath` as the base. Till now, our
settings is at default value (memory_quota - 256MB). This applies
new indexer setting of 256MB to indexer

5. next, settings manager processes the metakv update with new
memory_quota (eg 1.5GB) and tries to update the same in indexer

6. Now indexer sees that it has a different memory_quota so it
restarts

7. this restart of first indexer node addition to the cluster can cause
rebalance failures as indexer went down

hence we need to avoid this race

exp:
* avoid settings processing for empty value in metakv
* avoid races in metakvCallback and handleReady by adding locks
we need locks as we don't see this rebalance failures always
* add additional logging for future debugging of such issues

warn:
this works only when path IndexingSettingsShardAffinityMetaPath is
empty when first indexer node is added to the cluster. if this path has
a value and we see a metakv upgrade on IndexingSettingsMetaPath, we
still run into this restart. But this value should not exist without
any indexer node to modify it

Change-Id: Ib440f4be0c7596460017d596e1e6ef6dced45f21