MB-58481 Remove realInstId from book-keeping with emptyNodeBatching
When a proxy index is moving during rebalance, then indexer can
pick-up realInstId or proxy "instId" depending on the presence of
the index on the node. In a normal flow, the proxy will be merged
to realInstId. Hence, processedInst should always contain realInstId.

With empty node batching, since the merge is skipped, it is possible
that proxy is ready but rebalancer will not know whether indexer has
picked up proxy "instId" or realInstId.

Hence, delete the realInstId as well from the book-keeping so that
rebalancer will not wait forver.

Change-Id: Ifb9e91453647e842542112912dfa9d18601937fc