MB-58913 Skip waitQ throttling in recovery with empty node batching
In performance tests, it has been observed that recovery is the most
time consuming part. This is because shard rebalancer allows only
2 tokens to recover by default and among those two tokens, indexes
on one keyspace are recovered at a time. This is done to ensure that
indexer's resources are not saturated due to large number of
recovery operations triggered due to rebalance.

However, if empty node batching is enabled, this throttling can be
skipped as there wont be any scan traffic on the destination node.
This means that emptyNodeBatching will always be enabled for shard
rebalance irrespective of DCP tokens being present or not.

Change-Id: I0e9d86cfaf98ae838a21ab29d293248e56f215a6