Revert "MB-59052 Handle race between doFinish and Cancel()"
This reverts commit 3a3b1507989156657fd4f8ed9a4ecfb061671cf5.

Reason for revert: This code is built on the assumption that sync.Once is non blocking. This assumption is entirely wrong. If multiple go-routines execute sync.Once, then all the go-routines will be blocked until once invocation of sync.Once is executed. This will lead to a deadlock in the code.

a. doFinish() executes cleanupOnce which will be waiting for rebalance service manager lock in rebalanceDoneCallback

b. Cancel() while holding the rebalance service manager lock will wait for cleanupOnce to finish --> cleanupOnce will block as per the semantics and I was wrong in understanding the exact semantics.

Hence, reverting the changes.

Change-Id: I45dfe98a4d60c58683640a5381bdf092802aba64