MB-59052 Do not add sendPeerServerCommand to wait group
When Cancel() and doFinish race with each other, then the
waitGroup in doFinish::sendPeerServerCommand can race with
wg.Wait() in Cancel() and this can lead to a panic.

As this method is only sending a message to shard transfer
manager to start/stop the server (and not modifying any of
the token states), it is ok to call this method without
waitgroup synchronisation.

Change-Id: Ie0e0b88b330209666dff7eb417272d6625f7ac09