MB-58481 Use instId, RealInstId from transfer token for partition merge
For shard rebalance, the instId and realInstId are carried as a part of
tt.InstIds and tt.RealInstIds. They are not carried under the index
instance. Therefore, use tt.InstIds and tt.RealInstIds for partition
merge when empty node batching is enabled.

Also, do not acquire the ShardRebalancer lock when changing RState as
the token state or rebalancer book-keeping will remain intact. This
will help other tokens to merge instances and unblock merges which
are waiting for real instance RState move to Active.

Change-Id: Ide788304309016eb82939fa00ad558176f718034