MB-58908 Prune slots based on non-moving replicas
When pruning slots, planner also considers the replicas that are not
moving. This is done to handle the cases where there is cyclic
dependency in placement (or) a replica can not move to a node due to
the presence of another replica on the node. E.g.,

Cyclic dependency:
Node n1: i1 (replica 0), i2 (replica 0)
Node n2: i1 (replica 1), i2 (replica 2)
Node n3: i1 (replica 2), i2 (replica 1)

Node n1 is being swap rebalanced with n4 and planner wants to place
i1 and i2 on same slot.

If index "i2" were to be placed on a slot with "i1", then rebalance
would fail as i2 (replica 1) will be attempted to move to n2 while n2
already has i2 (replica 2). Unless both n2 and n3 are rebalanced out,
the conflict can not be resolved. If the cyclic dependencies exist
across multiple nodes, then it becomes more difficult to handle such
cases. To avoid such cases, also consider replicas that are not moving
to prune the slots further.

Change-Id: I74956e78cbf3db96a68c27b485fcb0659e7333c8