MB-59242 Initialize alterante shardIds in indexDefn after restart
After indexer restarts, it would read the topology from meta and
populate the index definition for each instance. However, the
alternate shardIds of the index metadata are not updated for the
specific index instance. This leads to indexer passing empty/wrong
alternate shardIds to plasma.

This patch fixes the issue by populating alternate shardIds specific
to the index instance after bootstrap.

Change-Id: If7e3c1d4aa99cbad468ac58d0b3c330e893d455a