MB-59646 Add support for nested flatten arrays
BP of MB-59138 to 7.2.4

Currently, projector does not explode null/missing vector entries
for flattened array indexes. When query returns [null ["a" 10]]
as the vector, projector has to explode the vector and transform
it to [[null null] ["a" 10]]. This patch adds support for the same.

Change-Id: I5d70c3569ed0b199a0ff438e71aac0c4f653ffb6
(cherry picked from commit 7b75cec57392c791627d740ff44762e020143001)