MB-59385 Transition to DCP rebalance when processing ShardTokenDeleted
Currently, transition to DCP rebalance happens only when the token
state changes to ShardTokenRecoveryShard with IsPendingReady set
to true for the token.

However, if there are few tokens with empty node batching enabled
and few with disbaled, then the following sequence of events will
lead to rebalance hang:

a. All tokens with empty node batching enabled are processed first
b. When IsPendingReady is set to true for these tokens, the check
   "allTokensPendingReadyOrDone" would fail as there are tokens
   with out empty node batching still waiting to be processed
c. Once all tokens without empty node batchig is disabled are
   processed, those tokens would move to ShardTokenDeleted state
d. At this point, there are a few tokens in ShardTokenRecoverShard
   phase with IsPendingReady set to true waiting for state change
   to ShardTokenMerge
e. As master is checking for only DeletedState for all the tokens,
   rebalance will endup in a hang

The fix for the issue is to:
a. Check for allTokensPendingReadyOrDone when empty node batching
   is enabled (insted of checkAllTokensDone)
b. Transition to DCP rebalance only if DCP rebalance is not already
   started. If DCP rebalance is already started in earlier phase,
   then finish rebalance as all tokens are processed.

Change-Id: I2f8e80ab9aa0781fc793870e20fc8098cbecd10b