MB-59239 Cleanup transferred data only if node is not active
When a node is active and other nodes initiate transfer cleanup,
then the node processing the token can end-up with in-consistent
state. E.g., consider the following:

a. Rebalance got cancelled and node n1 initiated cleanup
b. While n1 is processing cleanup, n2 is restoring shard from
   intermediate storage (E.g., S3) - n2 did not observe rebalance
   cancel yet
c. Node n1 cleaned up the transferred data from the intermediate
   storage while n2 is restoring from intermediate storage

In the CI test, this case resulted in plasma creating a new "shard0"
which is incorrect. To avoid such unwanted race conditions, cleanup
of transferred data happens only when the node is not "active" in the
cluster.

Change-Id: Id0233010030ce8ccc401b4d0c9a5af0eed1b99ad