MB-59460 Handle multiple proxies for same instance during bootstrap
During rebalance, it is possible that multiple partitions of an index
move to a single node. In such a case, each partition of the index will
have a separate proxy.

If indexer crashes before changing the RState of the proxy instances,
then the proxy partitions will be considered orphan and the partitions
will be cleaned up. This is because the "createRealInstIdMap" does
not account for the the presence of multiple proxies for a single index.

The fix for this issue is to merge partitions of multiple proxies in
realInstIdMap so that partition will not be treated orphan.

Change-Id: I8b329eb2878c7c62b0ea160564aa76ecdf5b6e12