MB-59460 Handle collection drop during rebalance
If a collection is dropped while an index state is in RebalPending,
then handleUpdateRState would fail (and also partition merge) as
the index would be deleted. This will lead to indexer panic.

The fix for this issue is to keep a track of all indexes whose
state is RebalPending when there is a collection drop and return
"ErrIndexDeletedDuringRebal" when the collection is dropped so
that rebalancer/shard rebalancer can handle the error and proceed
further.

Change-Id: Iab6bad6ba11f5369e748e14e1b9b671de6a1841e