MB-59460 Monitor for pruned partitions even if flush is in progress
Unlike dropIndex, partition prune would immediately return to caller
if flush is in progress. DropIndex would wait until flush is done.
During rebalance, if partition prune happens while flush is in progress,
rebalancer would assume that partition pruning was successful and initiate
shard destruction while data processing is still active (as shard tranfer
manager would not know what slices needs to be closed until prune is initiated)

To avoid such cases, indexer would send a message to shard transfer manager
to monitor for slices irrespective of flush so that at the time of pruning,
it is guaranteed that shard transfer manager knows about the slices that
needs to be closed before destroying a shard.

Change-Id: I3709369300f24e76946035c7f6a94634cda12b9d