MB-59465 Use all targetNodes as fullCapNodes when global limits are hit
When global limits on a shard are hit, planner tries to re-use the shard
slots instead of creating new shards inspite of indexer nodes having
capacity to create new shards.

In the current code, if fullCapNodes are empty, then planner goes ahead
and creates new shards. Instead, planner should consider all targetNodes
as fullCapNodes and then decide about generating new shards or not. If
planner can find a slot that is common to all targetNodes, then it will
be used. Otherwise, a new shard is created.

Change-Id: Ic4484399f90db1def03a1a1eb4d1329afa47c332