MB-59490 Use latest internal version when checking for shard affinity
The internalVersion in metadata provider is initialized when it is
created and never updated. Even for client auth, latest value of
internal version is retrieved and then used without updating provider's
internal version.

This patch also follows same pattern as client Auth. Ideally, provider's
internal version needs to be updated but the semantics are not clear
on how to handle metadata provider's internal version.

Change-Id: I894937622c178d10155dad460ad9f126c3681f2e