MB-59461 Skip slice closure for instances getting dropped during rebalance transfer
a. During rebalance, index instances can be dropped via collection
drop for non-serverless deployments. For serverless deployments, user
can initiate drop

b. When a shard transfer is in progress, indexes can not be dropped
as it will lead to corrupt shard data

c. For serverless, this is handled by keeping a track of rebalance
transfer for each bucket. If a bucket transfer is in progress, then
drop is skipped for any instance of that bucket

d. In non-serverless cases, as one shard can contain indexes beloning
to multiple buckets, keeping a track of per bucket rebalance will not
work. Therefore, rebalancer will now keep a track of per instance
transfer progress. Althgouh the book-keeping is per instance, the
status of all instances of a transfer token are updated together.

e. Indexer will use the instance transfer phases and build the bucket
transfer phase i.e. list of all buckets for which rebalance is in
progress. Lifecycle manager will use this information to allow (or)
dis-allow DDL during rebalance.

Change-Id: I8a44aa7c13029a15cc4a7d6fb0943cab4319e5cb